Refactor ImportSeed onboarding
