+++ packages/extension/src/components/pages/ImportSeed/Congratulations/index.ts
@@ -1,0 +1,1 @@
+export * from './Congratulations';
