+++ packages/extension/src/components/pages/ImportSeed/CreatePassword/index.ts
@@ -1,0 +1,1 @@
+export * from './CreatePassword';
