+++ packages/extension/src/components/pages/ImportSeed/constants/index.ts
@@ -1,0 +1,1 @@
+export * from './steps';
