+++ packages/extension/src/components/pages/ImportSeed/constants/steps.ts
@@ -1,0 +1,1 @@
+export const STEPS = 3;
